nycbuswatcher-siri-lambda/requirements.txt
requests
trio
asks
pandas
python-dateutil
fastparquet